refactor(#3038): VoiceBargeInRuntime S6 — extract TTS pipeline cluster
S6 scope: extract the next cohesive VoiceBargeInRuntime subcluster from the S1
concern map without behavior changes. The S5 closing assessment left the TTS
pipeline as the remaining mapped candidate (`set_runtime_tts_voice` 24,
`synthesize_acknowledgement` 9, `synthesize_progress_tts` 50 in the S1 map);
the other root residents are routing/dispatch (already split where clean by S4)
and small lifecycle/config + misc methods below a cohesive-cluster threshold.
The TTS pipeline moves cleanly as one ownership boundary — runtime TTS-voice
rebuild plus progress/ack synthesis — so it is the S6 pick.

Cluster boundary: moved `set_runtime_tts_voice`, `synthesize_acknowledgement`,
and `synthesize_progress_tts` verbatim into
`src/services/discord/voice_barge_in/tts_pipeline.rs` as an inherent impl child
module. The only moved-method signature edits are visibility-only:

- `set_runtime_tts_voice`: private -> `pub(super)` (root callers
  `apply_dispatcher_command` and `set_runtime_tts_voice_external`, plus the root
  test `runtime_config_snapshot_reflects_mutable_voice_state`).
- `synthesize_acknowledgement`: private -> `pub(super)` (root callers in
  `try_handle_voice_transcript_announcement`, `dispatch_voice_background_handoff`,
  and `process_completed_utterance`).
- `synthesize_progress_tts`: private -> `pub(super)` (its in-cluster caller
  `synthesize_acknowledgement` plus the sibling child caller
  `progress_playback::speak_progress_text`).

No root import mirror was needed: the moved bodies contain no `super::` paths and
every referenced type (`TtsRuntime`, `TtsSynthesisKind`, `PathBuf`, `ChannelId`,
`Ordering`) resolves through the child's `use super::*`. The root retains the
`crate::voice::tts::{TtsRuntime, TtsSynthesisKind}` import unchanged — `TtsRuntime`
still has root residents (the `tts` field type and `from_voice_config` in the
constructor) and `TtsSynthesisKind` is now consumed solely through the child's
glob re-export, so the root import stays warning-clean (verified under
`cargo clippy --lib -- -D warnings`).

No await point, lock scope, spawn, map-entry, cancellation, channel-ordering, or
queue-mutation behavior changed: the `voice_config_state`/`tts` RwLock write/read
scopes, the F10 (#2046) keep-previous-TTS error branch, and the `#[cfg(test)]`
synth-request capture block all moved byte-identical. No struct fields were
widened. Existing cfg behavior is unchanged.

LoC accounting: `voice_barge_in.rs` production LoC drops 3217 -> 3136 (-81). New
child `voice_barge_in/tts_pipeline.rs` is 86 production LoC, below the 1000-line
giant threshold, so only the root ratchet is lowered.

Token proof: normalized method-token comparison against HEAD is identical after
visibility-only normalization. Per moved method:

| Method | Method tokens | Body tokens | Normalized SHA |
| --- | ---: | ---: | --- |
| `set_runtime_tts_voice` | 170 -> 170 | 159 -> 159 | `65c9f41842d0` |
| `synthesize_acknowledgement` | 43 -> 43 | 20 -> 20 | `2ed0e4b90082` |
| `synthesize_progress_tts` | 289 -> 289 | 259 -> 259 | `34495fbecc21` |
| **Total** | **502 -> 502** | **438 -> 438** | |

REMAINING-MAP CLOSING ASSESSMENT (post-S6, guides whether S7 happens):

The S1 map's five named concerns are now distributed as follows. STT (S1),
progress playback (S2), final-result playback (S3), routing resolution (S4),
live-cut playback (S5), and the TTS pipeline (S6) have all been extracted into
sub-1000-LoC child modules. What remains in the root is:

- routing/dispatch HOT PATH: `try_handle_voice_channel_text_reply`,
  `handle_processing_transcript`, `try_handle_voice_transcript_announcement`
  (340), `dispatch_voice_background_handoff` (291), `start_voice_turn` (244),
  `process_completed_utterance` (163), `drain_deferred_after_turn`,
  `take_deferred_prompt`, `apply_dispatcher_command`, the inflight-foreground
  cancel registry (`register/unregister/cancel_inflight_foreground_cancel`),
  `register_voice_context`/`unregister_voice_guild`/`control_channel_ids_for_guild`,
  and `buffer_for_channel`. This is NOT a single clean ownership cluster: it is
  the orchestration core that threads SharedData, the deferred buffers, the
  cancel registry, the background driver handoff, and the now-extracted
  routing-resolution/playback/STT/TTS helpers together. The four largest methods
  (announcement 340, handoff 291, turn 244, completed-utterance 163) each call
  across three or more already-extracted clusters and into the struct's private
  field set; moving any subset verbatim would either pull mutually-recursive
  callers apart or force struct-field visibility widening (forbidden by the
  S-series discipline). There is no further clean verbatim slice here.

- lifecycle/config + misc: `from_voice_config`, `disabled`, `enabled`,
  `runtime_config_snapshot`, the verbose-progress/language/sensitivity setters,
  `apply_voice_command`, `apply_wake_word_command(_external)`,
  `cached_config`, `spawn_sensitivity_ttl_reset`, and the small
  ack/summary text generators. These total well under the cohesive-cluster
  threshold and are intentionally root-resident (constructor + the public
  command/config surface that callers reach through
  `pub(in crate::services::discord)`).

Conclusion: S6 closes the mapped extraction program. The five S1 concerns plus
the S4 routing-resolution boundary are all extracted. The dispatch core that
remains is the irreducible orchestration root, and the lifecycle/config residue
is below the slice threshold. NO further clean verbatim slice exists — S7 is NOT
warranted under the S-series discipline (verbatim-only, zero field-visibility
edits, child < 1000 LoC). Any further reduction would require behavior-touching
refactors (e.g. splitting the dispatch state into a sub-struct), which is out of
scope for this god-object decomposition program.

Gates run before commit:

- cargo fmt --check
- cargo check --lib
- cargo clippy --lib -- -D warnings
- set -o pipefail; cargo test --lib voice 2>&1 | grep "test result" (277
  passed) && cargo test --lib discord 2>&1 | grep "test result" (1620 passed)
- python3 scripts/audit_maintainability.py --check


Closing assessment (corrected per codex review P2): the TTS pipeline was the
largest remaining MAPPED candidate, but two further cohesive clusters remain
in root and are S7+ candidates: (a) the agent-voice routing helper block
(~agent_voice_matches_channel .. agent_text_channel_matches, lines ~398-485,
already consumed by voice_barge_in/routing.rs, no struct-field dependency),
and (b) the foreground decision/parser cluster starting ~line 599. The EPIC
is NOT closed by this slice.
