+++ scripts/audit_maintainability_giant_baseline.toml
@@ -109,7 +109,10 @@
 # #3038 VoiceBargeInRuntime S5: lowered 3332 -> 3217 (-115) as the live-cut
 # playback method cluster moved to `voice_barge_in/live_cut_playback.rs` (120
 # prod LoC, below the giant threshold). Locks in the shrink win.
-"src/services/discord/voice_barge_in.rs" = 3217
+# #3038 VoiceBargeInRuntime S6: lowered 3217 -> 3136 (-81) as the TTS pipeline
+# method cluster moved to `voice_barge_in/tts_pipeline.rs` (86 prod LoC, below
+# the giant threshold). Locks in the shrink win.
+"src/services/discord/voice_barge_in.rs" = 3136
 # #3248: +60 for the gap-1 deploy-survivable-relay fix —
 # `reseed_watcher_owned_finalizer_ledger` + two guarded reattach call-sites that
 # re-register the watcher-owned turn in the post-restart finalizer ledger (a P1
